[ANDROSDK-1982] move logic to network layer
