+++ core/src/main/java/org/hisp/dhis/android/core/relationship/RelationshipConstraintType.kt
@@ -25,9 +25,8 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package org.hisp.dhis.android.core.relationship
 
-package org.hisp.dhis.android.core.relationship;
-
-public enum RelationshipConstraintType {
+enum class RelationshipConstraintType {
     FROM, TO
 }
