+++ core/src/main/java/org/hisp/dhis/android/core/tracker/importer/internal/TrackerImporterPostCall.kt
@@ -47,6 +47,8 @@ import org.hisp.dhis.android.core.tracker.importer.internal.TrackerImporterObjec
 import org.hisp.dhis.android.core.tracker.importer.internal.TrackerImporterObjectType.EVENT
 import org.hisp.dhis.android.core.tracker.importer.internal.TrackerImporterObjectType.RELATIONSHIP
 import org.hisp.dhis.android.core.tracker.importer.internal.TrackerImporterObjectType.TRACKED_ENTITY
+import org.hisp.dhis.android.network.trackerimporter.IMPORT_STRATEGY_CREATE_AND_UPDATE
+import org.hisp.dhis.android.network.trackerimporter.IMPORT_STRATEGY_DELETE
 import org.koin.core.annotation.Singleton
 import java.util.Date
 
@@ -54,7 +56,7 @@ import java.util.Date
 internal class TrackerImporterPostCall internal constructor(
     private val payloadGenerator: NewTrackerImporterTrackedEntityPostPayloadGenerator,
     private val stateManager: NewTrackerImporterTrackedEntityPostStateManager,
-    private val service: TrackerImporterService,
+    private val networkHandler: TrackerImporterNetworkHandler,
     private val fileResourcesPostCall: TrackerImporterFileResourcesPostCall,
     private val programOwnerPostCall: TrackerImporterProgramOwnerPostCall,
     private val coroutineAPICallExecutor: CoroutineAPICallExecutor,
@@ -124,7 +126,7 @@ internal class TrackerImporterPostCall internal constructor(
         importStrategy: String,
     ): Result<String, D2Error> {
         return coroutineAPICallExecutor.wrap(storeError = true) {
-            service.postTrackerPayload(payload, ATOMIC_MODE_OBJECT, importStrategy)
+            networkHandler.postTrackerPayload(payload, importStrategy)
         }.map { res ->
             val jobId = res.response().uid()
             jobObjectHandler.handleMany(generateJobObjects(payload, jobId))
